Feat food

See merge request s09-final/S09P31B205!30

Former-commit-id: 1e7c6539058f98c74f812369fa47accd56113434

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/hotel/dto/request/AttractionNameRequestDto.java
@@ -8,6 +8,6 @@ import lombok.*;
 @AllArgsConstructor
 public class AttractionNameRequestDto {
     private String attractionName;
-    private Long distance;
+    private int distance;
     private String sorted;
 }
